Fix up OTG mentions and Pi model A reference
Just like most of the Raspberry Pi community, the OTG acronym is used the other way around here. [USB On-The-Go](https://en.wikipedia.org/wiki/USB_On-The-Go) is a standard that allows USB *devices* to act as USB *hosts*. Raspberry Pi default bahaviour is to have OTG enabled so it acts like a *host* and you can attach your own *peripherals*. The purpose of this repository is to disable the OTG in order to make Raspberry Pi act like a *device* so it can provide webcam function to another *host*.

Also, for Raspberry Pi <= 3 model A, the USB data wires that can be used in the *device* mode are wired to the type-A connector, not to the power micro-USB port. So you have to use a non-standard cable with type-A connectors on both ends. For Pi <= 3 model B, the device mode is impossible without soldering, because there is a USB hub onboard.